change pc and eh representation to u16

+++ src/instruction_handlers/far_call.rs
@@ -38,7 +38,7 @@ fn far_call<const CALLING_MODE: u8, const IS_STATIC: bool>(
 
     let raw_abi = Register1::get(args, &mut vm.state);
     let destination_address = Register2::get(args, &mut vm.state) & address_mask;
-    let exception_handler = Immediate1::get(args, &mut vm.state).low_u32();
+    let exception_handler = Immediate1::get(args, &mut vm.state).low_u32() as u16;
 
     let abi = get_far_call_arguments(raw_abi);
